requirements.txt
torch
torchvision
ninja
numpy
pillow
uvicorn
fastapi
aiofiles
python-multipart